Add shell script to start Docker container for development environment
This commit adds a short shell script to quickly start and run a new
development environment of the site. Making this available in the
repository is a time-saving step for someone who wants to quickly test
out a change by running a simple script to start everything up for them.